+++ adapters/sqlite3/sqlite3.go
@@ -1,6 +1,6 @@
 package sqlite3
 
-import "gopkg.in/doug-martin/goqu.v4"
+import "github.com/doug-martin/goqu/v6"
 
 var (
 	placeholder_rune    = '?'