+++ dataset_actions.go
@@ -141,7 +141,7 @@ func (me *Dataset) Insert(i ...interface{}) *CrudExec {
 }
 
 //Generates the INSERT IGNORE (mysql) or INSERT ... ON CONFLICT DO NOTHING (postgres) and returns an Exec struct.
-//    db.From("test").InsertIgnore(DoNothing(), Record{"name":"Bob").Exec()
+//    db.From("test").InsertIgnore(DoNothing(), Record{"name":"Bob"}).Exec()
 //
 //See Dataset#InsertIgnore for arguments
 func (me *Dataset) InsertIgnore(i ...interface{}) *CrudExec {
@@ -150,7 +150,7 @@ func (me *Dataset) InsertIgnore(i ...interface{}) *CrudExec {
 }
 
 //Generates the INSERT sql with (ON CONFLICT/ON DUPLICATE KEY) clause, and returns an Exec struct with the sql set to the INSERT statement
-//    db.From("test").InsertConflict(DoNothing(), Record{"name":"Bob").Exec()
+//    db.From("test").InsertConflict(DoNothing(), Record{"name":"Bob"}).Exec()
 //
 //See Dataset#Upsert for arguments
 func (me *Dataset) InsertConflict(c ConflictExpression, i ...interface{}) *CrudExec {