+++ example_test.go
@@ -4,9 +4,11 @@ import (
 	"database/sql"
 	"fmt"
 	"regexp"
+	"strings"
 
-	"gopkg.in/DATA-DOG/go-sqlmock.v1"
-	"gopkg.in/doug-martin/goqu.v4"
+	"github.com/doug-martin/goqu/v6"
+
+	"github.com/DATA-DOG/go-sqlmock"
 )
 
 var driver *sql.DB
@@ -614,8 +616,8 @@ func ExampleDataset_WithCTERecursive() {
 	sql, _, _ := db.From("nums").
 		WithRecursive("nums(x)",
 			db.From().Select(goqu.L("1")).
-			UnionAll(db.From("nums").
-				Select(goqu.L("x+1")).Where(goqu.I("x").Lt(5)))).
+				UnionAll(db.From("nums").
+					Select(goqu.L("x+1")).Where(goqu.I("x").Lt(5)))).
 		ToSql()
 	fmt.Println(sql)
 	// Output:
@@ -1857,4 +1859,30 @@ func ExampleDataset_Prepared() {
 	// UPDATE "items" SET "address"=?,"name"=? [111 Test Addr Test]
 	// DELETE FROM "items" WHERE ("id" > ?) [10]
 
+}
+
+func ExampleSetColumnRenameFunction() {
+	mDb, mock, _ := sqlmock.New()
+
+	mock.ExpectQuery(`SELECT "ADDRESS", "NAME" FROM "items" LIMIT 1`).
+		WithArgs().
+		WillReturnRows(sqlmock.NewRows([]string{"ADDRESS", "NAME"}).FromCSVString("111 Test Addr,Test1"))
+
+	db := goqu.New("db-mock", mDb)
+
+	goqu.SetColumnRenameFunction(strings.ToUpper)
+
+	anonStruct := struct {
+		Address string
+		Name    string
+	}{}
+	found, _ := db.From("items").ScanStruct(&anonStruct)
+	fmt.Println(found)
+	fmt.Println(anonStruct.Address)
+	fmt.Println(anonStruct.Name)
+
+	// Output:
+	// true
+	// 111 Test Addr
+	// Test1
 }